renamed modules to streamed collections
