+++ rocketchat-core/src/main/java/com/rocketchat/core/ChatRoom.java
@@ -11,9 +11,9 @@ import com.rocketchat.core.callback.FileListener;
 import com.rocketchat.core.callback.HistoryCallback;
 import com.rocketchat.core.callback.MessageCallback;
 import com.rocketchat.core.callback.RoomCallback;
-import com.rocketchat.core.db.Document.FileDocument;
-import com.rocketchat.core.db.Document.MessageDocument;
-import com.rocketchat.core.db.LocalStreamCollectionManager;
+import com.rocketchat.core.roomstream.Document.FileDocument;
+import com.rocketchat.core.roomstream.Document.MessageDocument;
+import com.rocketchat.core.roomstream.LocalStreamCollectionManager;
 import com.rocketchat.core.internal.middleware.CoreStreamMiddleware;
 import com.rocketchat.core.model.FileDescriptor;
 import com.rocketchat.core.model.Message;
