+++ sample/src/main/java/com/rocketchat/sample/Main.java
@@ -1,21 +1,15 @@
 package com.rocketchat.sample;
 
 import com.rocketchat.common.RocketChatException;
-import com.rocketchat.common.data.lightdb.document.ClientVersionsDocument;
-import com.rocketchat.common.data.lightdb.document.LoginConfDocument;
-import com.rocketchat.common.data.lightdb.document.RocketChatRolesDocument;
-import com.rocketchat.common.data.lightdb.document.UserDocument;
 import com.rocketchat.common.listener.ConnectListener;
 import com.rocketchat.common.listener.SimpleListCallback;
 import com.rocketchat.common.listener.StreamCollectionListener;
 import com.rocketchat.common.listener.SubscribeListener;
-import com.rocketchat.common.network.ReconnectionStrategy;
 import com.rocketchat.common.utils.Logger;
 import com.rocketchat.core.ChatRoom;
 import com.rocketchat.core.RocketChatClient;
 import com.rocketchat.core.callback.LoginCallback;
-import com.rocketchat.core.db.Document.FileDocument;
-import com.rocketchat.core.db.Document.MessageDocument;
+import com.rocketchat.core.roomstream.Document.MessageDocument;
 import com.rocketchat.core.model.Subscription;
 import com.rocketchat.core.model.Token;
 import java.util.List;
